book update

+++ app/Repository/BookRepository.php
@@ -59,26 +59,14 @@ class BookRepository implements IBookRepository
       $book->publisher_id = $attributes['publisher'];
       $book->description = $attributes['description'];
 
-      // $data = [
-      //   'id' => $bookId,
-      //   'title' => ucwords($attributes['title']),
-      //   'num_pages' => $attributes['numPages'],
-      //   'author_id' => $attributes['author'],
-      //   'publisher_id' => $attributes['publisher'],
-      //   'description' => $attributes['description'],
-      // ];
-    
       // Add Book Cover
       $book->slug = Str::slug($book->title);
-
-      $bookCoverUrl = $book->slug . '-' . time() . '.' . $attributes['cover']->extension();
-
-      $book->cover_url = 'bookCovers/' . $bookCoverUrl;
+      $book->cover_url = 'bookCovers/' . $book->slug . '-' . time() . '.' . $attributes['cover']->extension();
       
-      Storage::disk('public')->putFileAs('bookCovers', $attributes['cover'], $bookCoverUrl);
-      array_push($fileStored, $book['cover_url']);
+      Storage::disk('public')->put($book->cover_url, file_get_contents($attributes['cover']));
+      array_push($fileStored, $book->cover_url);
 
-      // Storage::disk('public')->put('bookCovers/'.$bookCoverUrl, $attributes['cover']);
+      $book->save();
 
       // Add book genre
       if (isset($attributes['genres'])) {
@@ -91,31 +79,33 @@ class BookRepository implements IBookRepository
       $fileTypes = FileType::all();
       foreach ($fileTypes as $fileType) {
         if (isset($attributes[$fileType->name])) {
-          $bookFileUrl = $fileType->name . '/' . $book->slug . '-' . time() . '.' . $attributes[$fileType->name]->extension();
-          array_push($fileStored, $bookFileUrl);
+          $bookFileUrl = 'files/' . $fileType->name . '/' . $book->slug . '-' . time() . '.' . $attributes[$fileType->name]->extension();
 
           BookFile::create([
             'book_id' => $bookId,
             'file_type_id' => $fileType->id,
-            'file_url' => 'files/'.$bookFileUrl
+            'file_url' => $bookFileUrl
           ]);
   
-          Storage::put('files/'.$bookFileUrl, $attributes[$fileType->name]);
+          Storage::put($bookFileUrl, file_get_contents($attributes[$fileType->name]));
+
+          array_push($fileStored, $bookFileUrl);
+
         }
       }
   
-      $book->save();
-
       DB::commit();
       
       return true;
 
     } catch (\Throwable $th) {
-      dd($th);
       DB::rollBack();
 
-      foreach ($fileStored as $file) {
-        Storage::delete($file);
+      foreach ($fileStored as $filename) {
+
+        if(in_array(pathinfo($filename, PATHINFO_EXTENSION), ['png', 'jpg'])) Storage::disk('public')->delete($filename);
+        else Storage::delete($filename);
+
       }
 
       return false;
@@ -124,15 +114,15 @@ class BookRepository implements IBookRepository
   }
 
   public function update($book = null, $attributes = null) {
-    
     $data = [];
 
-    $fileStored = [];
-    $fileDeleted = [];
+    $storedFiles = [];
+    $deletedFiles = [];
 
     $storedCoverUrl = '';
     $deletedCoverUrl = '';
 
+    DB::beginTransaction();
     try {
       if($book->slug != Str::slug($attributes['title'])) {
         $data['title'] = ucwords($attributes['title']);
@@ -150,31 +140,78 @@ class BookRepository implements IBookRepository
           ? 'bookCovers/' . $data['slug'] . '-' . time() . '.' . $attributes['cover']->extension()
           : 'bookCovers/' . $book->slug . '-' . time() . '.' . $attributes['cover']->extension();
 
-        // $storedCover = $data['cover_url'];
-        // $deletedCover = 'deletedFiles/'.basename($currentBookUrl);
-
-        Storage::disk('public')->putFileAs('bookCovers/', $attributes['cover'], basename($data['cover_url']));
-
+        // Store new book cover to public disk
+        Storage::disk('public')->put('bookCovers/'.basename($data['cover_url']), file_get_contents($attributes['cover']));
         $storedCoverUrl = $data['cover_url'];
+
+        // Move current book cover to deletedFiles folder of local disk
+        Storage::put('deletedFiles/'.basename($currentBookUrl), Storage::disk('public')->get($currentBookUrl));
         $deletedCoverUrl = 'deletedFiles/'.basename($currentBookUrl);
 
-        $deleteBookCover = Storage::disk('public')->get($currentBookUrl);
-        Storage::put('deletedFiles/'.basename($currentBookUrl) , $deleteBookCover);
+        // Delete current book cover in public disk
         Storage::disk('public')->delete($currentBookUrl);
       }
       
-      throw new Exception("Error Processing Request", 1);
-      
-      $book->update($data);
+      $fileTypes = FileType::all();
+      foreach($fileTypes as $fileType) {
+        if(isset($attributes[$fileType->name])) {
+          $bookFile = BookFile::where([ ['book_id', '=', $book->id], ['file_type_id', '=', $fileType->id] ])->first();
+
+          if($bookFile) {
+            $newFileUrl = 'files/' . $fileType->name . '/' . $book->slug . '-' . time() . '.' . $attributes[$fileType->name]->extension();
+
+            Storage::put($newFileUrl, file_get_contents($attributes[$fileType->name]));
+            array_push($storedFiles, $newFileUrl);
+
+            Storage::put('deletedFiles/'. $bookFile->file_url, Storage::get($bookFile->file_url));
+            array_push($deletedFiles, 'deletedFiles/'. $bookFile->file_url);
+
+            Storage::delete($bookFile->file_url);
+
+            $bookFile->update(['file_url' => $newFileUrl]);
+          }
+          else {
+            $newFileUrl = 'files/' . $fileType->name . '/' . $book->slug . '-' . time() . '.' . $attributes[$fileType->name]->extension();
+
+            BookFile::create([
+              'book_id' => $book->id,
+              'file_type_id' => $fileType->id,
+              'file_url' => $newFileUrl
+            ]);
+    
+            Storage::put($newFileUrl, file_get_contents($attributes[$fileType->name]));
+
+            array_push($storedFiles, $newFileUrl);
+          }
+        }
+      }
+
+      if($data) $book->update($data);
+
+      // dd($deletedFiles, $storedFiles);
+      // throw new Exception("Error Processing Request", 1);
+
+      DB::commit();
 
       return true;
 
     } catch (\Throwable $th) {
-      $reliveBookCover = Storage::get('deletedFile/'.$deletedCoverUrl);
-      
-      Storage::disk('public')->putFileAs('bookCovers', $reliveBookCover, basename($deletedCoverUrl));
-      Storage::delete($deletedCoverUrl);
-      Storage::disk('public')->delete($storedCoverUrl);
+      DB::rollBack();
+
+      if($storedCoverUrl && $deletedCoverUrl) {
+        Storage::disk('public')->put('bookCovers/'.basename($deletedCoverUrl), Storage::get($deletedCoverUrl));
+        Storage::delete($deletedCoverUrl);
+        Storage::disk('public')->delete($storedCoverUrl);
+      }
+
+      foreach($deletedFiles as $filename) {
+        Storage::put('files/'. explode('/', $filename)[2] . '/' .basename($filename), Storage::get($filename));
+        Storage::delete($filename);
+      }
+
+      foreach ($storedFiles as $filename) {
+        Storage::delete($filename);
+      }
 
       return false;
     }
